make settings work
