+++ admin/src/components/Wysiwyg/index.js
@@ -26,104 +26,32 @@ import TableCellExtension from '@tiptap/extension-table-cell'
 import TableHeaderExtension from '@tiptap/extension-table-header'
 import packageInfo from '../../../../package.json'
 
-const CSSColumnsExtension = Extension.create({
-  name: 'cssColumns',
-  addOptions () {
-    return {
-      types: [],
-      columnTypes: [2, 3],
-      defaultColumnType: 'two',
-    };
-  },
-  addGlobalAttributes () {
-    return [
-      {
-        types: this.options.types,
-        attributes: {
-          cssColumns: {
-            default: 1,
-            renderHTML: attributes => {
-              return {
-                style: `column-count: ${attributes.cssColumns}`,
-              }
-            },
-            parseHTML: element => element.style.columnCount || 1,
-          },
-        },
-      }
-    ]
-  },
-  addCommands () {
-    return {
-      toggleColumns: (columnType) => ({ commands, editor }) => {
-        if (!editor.isActive({'cssColumns': columnType})) return this.options.types.every((type) => commands.updateAttributes(type, {cssColumns: columnType}))
-        return this.options.types.every((type) => commands.resetAttributes(type, 'cssColumns'))
-      },
-      unsetColumns: (columnType) => ({ commands }) => {
-        return this.options.types.every((type) => commands.resetAttributes(type, 'cssColumns'))
-      },
-    }
-  }
-})
+
 
 
 const Wysiwyg = ({ name, onChange, value, intlLabel, disabled, error, description, required }) => {
   const { formatMessage } = useIntl();
   const [mediaLibVisible, setMediaLibVisible] = useState(false);
   const [forceInsert, setForceInsert] = useState(false);
-  const {data: settings, isLoading} = useQuery('settings', getSettings)
+  const [ mergedSettings, setMergedSettings] = useState(null)
+  const {data: settings, isLoading} = useQuery('settings', getSettings, {
+    onSuccess: async (values) => {
+      console.log('settings on success', values)
+      setMergedSettings({...defaultSettings, ...values})
+    },})
 
-  const mergedSettings = {...defaultSettings, ...settings}
 
   const handleToggleMediaLib = () => setMediaLibVisible(prev => !prev);
 
-  console.log(mergedSettings)
-
-  const editor = useEditor({
-    extensions: [
-      StarterKit.configure({
-        gapcursor: true,
-      }),
-      UnderlineExtension,
-      mergedSettings.links.enabled ? LinkExtension.configure({
-        autolink: mergedSettings.links.autolink,
-        openOnClick: mergedSettings.links.openOnClick,
-        linkOnPaste: mergedSettings.links.linkOnPaste,
-      }) : null,
-      mergedSettings.image.enabled ? ImageExtension.configure({
-        inline: mergedSettings.image.inline,
-        allowBase64: mergedSettings.image.allowBase64,
-      }) : null,
-      TextAlignExtension.configure({
-        types: ['heading', 'paragraph'],
-      }),
-      TableExtension.configure({
-        allowTableNodeSelection: true,
-      }),
-      TableRowExtension,
-      TableCellExtension,
-      TableHeaderExtension,
-      CSSColumnsExtension.configure({
-        types: ['paragraph']
-      }),
-    ],
-    onUpdate(ctx) {
-      console.log('Change')
-      onChange({ target: { name, value: ctx.editor.getHTML() } })
-    },
-  })
-
-
-
   // Update content if value is changed outside
-  useEffect(() => {
+  /*useEffect(() => {
     if (editor !== null && editor.getHTML() !== value && value !== null) {
       editor.commands.setContent(value)
     }
 
     if (!isLoading) {
     }
-  })
+  })*/
 
   const handleChangeAssets = assets => {
     if (!forceInsert && editor.isActive('image')) {
@@ -145,7 +73,11 @@ const Wysiwyg = ({ name, onChange, value, intlLabel, disabled, error, descriptio
   };
 
   if (isLoading) {
+    console.log('isLoading')
     return null
+  } else {
+    console.log('Is not loading')
+
   }
 
   return (
@@ -161,11 +93,11 @@ const Wysiwyg = ({ name, onChange, value, intlLabel, disabled, error, descriptio
           </Box>
           {/*<Button startIcon={<Landscape />} variant='secondary' fullWidth onClick={() => {setForceInsert(true); handleToggleMediaLib()}}>Media library</Button>*/}
           <Editor
+              key={isLoading ? 'loading' : 'loaded'}
               disabled={disabled}
               name={name}
               onChange={onChange}
               value={value}
-              editor={editor}
               toggleMediaLib={handleToggleMediaLib}
               settings={mergedSettings}
           />
